message fix
