+++ include/aquarius/router/router.hpp
@@ -1,12 +1,11 @@
 #pragma once
+#include <aquarius/context.hpp>
+#include <aquarius/context/generator.hpp>
 #include <aquarius/context/invoke.hpp>
 #include <aquarius/core/core.hpp>
 #include <aquarius/core/elastic.hpp>
 #include <aquarius/core/logger.hpp>
-#include <aquarius/message/invoke.hpp>
 #include <aquarius/message/message.hpp>
-#include <aquarius/context.hpp>
-#include <aquarius/context/generator.hpp>
 
 namespace aquarius
 {
